requirements.txt
a2wsgi==1.10.8
absl-py==2.3.0
aiofile==3.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiopath==0.6.11
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==3.7.1
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-sqlite==3.8.2
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
Brotli==1.1.0
bs4==0.0.2
cachelib==0.13.0
caio==0.9.24
catalogue==2.0.10
cattrs==25.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cmdstanpy==1.2.5
colorama==0.4.6
colorlog==4.8.0
confection==0.1.5
ConfigUpdater==3.2
connexion==3.2.0
contourpy==1.3.2
craiyon.py==1.2.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.3
cssselect==1.3.0
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
email-validator==1.3.1
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.3.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.58.1
fpdf==1.7.2
frozenlist==1.6.0
gast==0.6.0
google-pasta==0.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.70.0
graphviz==0.20.3
groq==0.26.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
holidays==0.73
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.11.0
libclang==18.1.1
lightgbm==4.6.0
limits==5.2.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.4.0
mailchimp-marketing==3.0.80
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.3
mdit-py-plugins==0.4.2
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.4.4
murmurhash==1.0.13
namex==0.1.0
newspaper3k==0.2.8
nltk==3.9.1
notion-client==2.3.0
numpy==2.1.3
oauthlib==3.2.2
openai==1.82.1
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opt_einsum==3.4.0
optree==0.16.0
ordered-set==4.1.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pillow==11.2.1
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
prison==0.2.1
propcache==0.3.1
prophet==1.1.7
protobuf==4.25.8
psutil==7.0.0
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
PySocks==1.7.1
pytest==8.3.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
setproctitle==1.3.6
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
srsly==2.5.1
stanio==0.5.1
starlette==0.46.2
svgwrite==1.4.3
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
Tree==0.2.4
tweepy==4.15.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xgboost==3.0.2
yarl==1.20.0
zopfli==0.2.3.post1

flask
